backend/requirements.txt
# Core frameworks
fastapi==0.115.0
flask==3.0.3
uvicorn==0.30.6
gunicorn==22.0.0

# Async and performance
anyio==4.4.0
uvloop==0.19.0
httptools==0.7.1
websockets==12.0
watchfiles==1.1.1

# Data processing and ML
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.2
xgboost==3.1.0
joblib==1.4.2
scipy==1.13.1

# AWS / S3 integration
boto3==1.35.45
botocore==1.35.45
s3transfer==0.10.2

# Database
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9

# API connectors and networking
requests==2.32.3
tldextract==5.1.2
shodan==1.31.0
virustotal-python==1.2.0
aiohttp==3.10.10
aiodns==3.2.0
aiofiles==24.1.0

# NVD + Security intelligence
nvdlib==0.7.6
abuseipdb==0.3.0   # if you use the API directly
certifi==2025.10.5

# Environment and configuration
python-dotenv==1.0.1
PyYAML==6.0.2
typing_extensions==4.12.2

# PDF / report generation
fpdf==1.7.2
xlsxwriter==3.2.0

# Logging, utility & color
colorama==0.4.6
click==8.1.7
click-plugins==1.1.1

# Date/time handling
python-dateutil==2.9.0.post0
pytz==2024.1
tzdata==2025.2

# Miscellaneous
idna==3.7
charset-normalizer==3.4.0
urllib3==2.2.2
six==1.16.0
greenlet==3.0.3
sniffio==1.3.1
jmespath==1.0.1
